service/requirements.txt
requests==2.22.0
confluence-rest-library
jinja2